+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #56 - Ability to precompile select Numba/CuPy kernel on import
 - PR #60 - Updated decimate function to use an existing FIR window
 - PR #65 - Added deprecation warning for Numba kernels
+- PR #67 - cuSignal code refactor and documentation update
 
 ## Bug Fixes
 - PR #44 - Fix issues in pytests